# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RAG Chat API",
    "description": "Multimodal RAG Platform API (PostgreSQL 16 + pgvector + Hybrid Search)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reports/overview": {
      "get": {
        "tags": [
          "Reports & Analytics (Composition)"
        ],
        "summary": "Báo cáo tổng quan không gian làm việc (Multi-table Overview)",
        "description": "Tổng hợp chỉ số đa bảng (Workspace, Members, Documents, Chunks, Sessions, Messages, Tokens, Feedback).",
        "operationId": "get_workspace_overview_report_api_v1_reports_overview_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc.",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOverviewReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "reports:read"
        ]
      }
    },
    "/api/v1/reports/activity": {
      "get": {
        "tags": [
          "Reports & Analytics (Composition)"
        ],
        "summary": "Báo cáo xu hướng hoạt động theo ngày",
        "description": "Thống kê số lượng upload tài liệu, tin nhắn và phiên hội thoại mới trong N ngày gần nhất.",
        "operationId": "get_workspace_daily_activity_report_api_v1_reports_activity_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc.",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc."
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "description": "Số ngày cần thống kê (1 - 90 ngày)",
              "default": 30,
              "title": "Days"
            },
            "description": "Số ngày cần thống kê (1 - 90 ngày)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDailyActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "reports:read"
        ]
      }
    },
    "/reports/overview": {
      "get": {
        "tags": [
          "Reports & Analytics (Composition)"
        ],
        "summary": "Báo cáo tổng quan không gian làm việc (Multi-table Overview)",
        "description": "Tổng hợp chỉ số đa bảng (Workspace, Members, Documents, Chunks, Sessions, Messages, Tokens, Feedback).",
        "operationId": "get_workspace_overview_report_reports_overview_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc.",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOverviewReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "reports:read"
        ]
      }
    },
    "/reports/activity": {
      "get": {
        "tags": [
          "Reports & Analytics (Composition)"
        ],
        "summary": "Báo cáo xu hướng hoạt động theo ngày",
        "description": "Thống kê số lượng upload tài liệu, tin nhắn và phiên hội thoại mới trong N ngày gần nhất.",
        "operationId": "get_workspace_daily_activity_report_reports_activity_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc.",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn Workspace ID cần xem báo cáo. Mặc định là active workspace của phiên làm việc."
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "description": "Số ngày cần thống kê (1 - 90 ngày)",
              "default": 30,
              "title": "Days"
            },
            "description": "Số ngày cần thống kê (1 - 90 ngày)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDailyActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "reports:read"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChatStatsDTO": {
        "properties": {
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions",
            "default": 0
          },
          "total_messages": {
            "type": "integer",
            "title": "Total Messages",
            "default": 0
          },
          "user_messages": {
            "type": "integer",
            "title": "User Messages",
            "default": 0
          },
          "assistant_messages": {
            "type": "integer",
            "title": "Assistant Messages",
            "default": 0
          },
          "total_prompt_tokens": {
            "type": "integer",
            "title": "Total Prompt Tokens",
            "default": 0
          },
          "total_completion_tokens": {
            "type": "integer",
            "title": "Total Completion Tokens",
            "default": 0
          },
          "total_citations": {
            "type": "integer",
            "title": "Total Citations",
            "default": 0
          },
          "average_latency_ms": {
            "type": "number",
            "title": "Average Latency Ms",
            "default": 0.0
          },
          "total_feedbacks": {
            "type": "integer",
            "title": "Total Feedbacks",
            "default": 0
          },
          "positive_feedbacks": {
            "type": "integer",
            "title": "Positive Feedbacks",
            "default": 0
          },
          "negative_feedbacks": {
            "type": "integer",
            "title": "Negative Feedbacks",
            "default": 0
          }
        },
        "type": "object",
        "title": "ChatStatsDTO",
        "description": "Aggregated chat, token usage, latency, and feedback metrics."
      },
...
      "DailyActivityItemDTO": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date",
            "description": "ISO date string (YYYY-MM-DD)"
          },
          "documents_uploaded": {
            "type": "integer",
            "title": "Documents Uploaded",
            "default": 0
          },
          "messages_sent": {
            "type": "integer",
            "title": "Messages Sent",
            "default": 0
          },
          "sessions_created": {
            "type": "integer",
            "title": "Sessions Created",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "DailyActivityItemDTO",
        "description": "Daily activity point for trend charts."
      },
...
      "DocumentStatsDTO": {
        "properties": {
          "total_documents": {
            "type": "integer",
            "title": "Total Documents",
            "default": 0
          },
          "total_file_size_bytes": {
            "type": "integer",
            "title": "Total File Size Bytes",
            "default": 0
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "default": 0
          },
          "ready_documents": {
            "type": "integer",
            "title": "Ready Documents",
            "default": 0
          },
          "failed_documents": {
            "type": "integer",
            "title": "Failed Documents",
            "default": 0
          },
          "processing_documents": {
            "type": "integer",
            "title": "Processing Documents",
            "default": 0
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks",
            "default": 0
          }
        },
        "type": "object",
        "title": "DocumentStatsDTO",
        "description": "Aggregated document metrics within a workspace."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberStatsDTO": {
        "properties": {
          "total_members": {
            "type": "integer",
            "title": "Total Members",
            "default": 0
          }
        },
        "type": "object",
        "title": "MemberStatsDTO",
        "description": "Workspace membership summary."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceDailyActivityResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityItemDTO"
            },
            "type": "array",
            "title": "Activities"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "days",
          "activities"
        ],
        "title": "WorkspaceDailyActivityResponse",
        "description": "Time-series activity report for charting and trends."
      },
...
      "WorkspaceOverviewReportResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "workspace_name": {
            "type": "string",
            "title": "Workspace Name"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "members": {
            "$ref": "#/components/schemas/MemberStatsDTO"
          },
          "documents": {
            "$ref": "#/components/schemas/DocumentStatsDTO"
          },
          "chat": {
            "$ref": "#/components/schemas/ChatStatsDTO"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "workspace_name",
          "generated_at",
          "members",
          "documents",
          "chat"
        ],
        "title": "WorkspaceOverviewReportResponse",
        "description": "Comprehensive multi-table overview report for a workspace."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token",
        "description": "Session token stored in HTTP-only cookie"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token",
        "description": "Session token cookie matching ViShop's @ApiCookieAuth('session')"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token passed in Authorization: Bearer <token>"
      }
    }
  }
}